docs/requirements.txt
# Documentation dependencies for ReadTheDocs
# Version: 2.0.0

# Core Sphinx and theme
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.25.0
myst-parser>=2.0.0

# Essential Sphinx extensions
sphinx-copybutton>=0.5.2

# Optional extensions (will be loaded if available)
# sphinx-design>=0.5.0
# sphinxcontrib-mermaid>=0.9.2
# nbsphinx>=0.9.1

# Standard Sphinx contrib extensions
sphinxcontrib-applehelp>=1.0.4
sphinxcontrib-devhelp>=1.0.2
sphinxcontrib-htmlhelp>=2.0.1
sphinxcontrib-serializinghtml>=1.1.5
sphinxcontrib-qthelp>=1.0.3

# Core dependencies needed for documentation build
numpy>=1.23.5,<2.0.0
scipy>=1.10.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
opencv-python>=4.8.1,<5.0.0
Pillow>=10.0.0,<12.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scikit-image>=0.20.0,<1.0.0
tqdm>=4.66.0,<5.0.0
psutil>=5.9.0,<6.0.0
